# DROP DB Performance improvements