### Description

Similar fixes to [Create DB PR](https://github.com/babelfish-for-postgresql/babelfish_extensions/pull/1899)
Relies on this as well [PR LINK](https://github.com/babelfish-for-postgresql/postgresql_modified_for_babelfish/pull/234)

For DROP DB we currently,
-- assume role sysadmin
-- drop all related objects (roles, schema, etc)
-- reset role to prev user


1. DROP SCHEMA -  the owner can drop the schema (and thereby all contained objects) even if they do not own some of the objects within the schema. [Reference Link](https://www.postgresql.org/docs/current/sql-dropschema.html)
2. DROP OWNED BY role1, role2 - drops all the objects within the current database that are owned by one of the specified roles. Any privileges granted to the given roles on objects in the current database or on shared objects (databases, tablespaces, configuration parameters) will also be revoked. [Reference Link](https://www.postgresql.org/docs/current/sql-drop-owned.html)

Based on the above definitions we can set the current_user to dbo role and still be able to drop everything inside the babelfish database with the only exception being the PG Database privileges granted to any roles being dropped. Since only sysadmin has the grant rights to PG Database, we add an additional revoke statement for dropping these privileges under the assumed role sysadmin.

This improves the performance by reducing the time taken to run roles_is_member_of( roleId )
When no of DBs >> 1
  roles_is_member_of( 'any_db_name_dbo' ) << roles_is_member_of( 'sysadmin' )

So we set the current user to dbo wherever possible.

| DROP DB              | After | Before |
| :----------------: | :------: | :----: |
| DBs2990   |   15 ms  | NA |
| DBs1000   |    7 ms  | 329 ms |
| DBs500     |   6 ms  | 98 ms |
| DBs10        |   4 ms  | 6 ms |

### Issues Resolved

[BABEL-4438]

Signed-off-by: Tanzeel Khan <tzlkhan@amazon.com>

### Check List
- [x] Commits are signed per the DCO using --signoff 

By submitting this pull request, I confirm that my contribution is under the terms of the Apache 2.0 and PostgreSQL licenses, and grant any person obtaining a copy of the contribution permission to relicense all or a portion of my contribution to the PostgreSQL License solely to contribute all or a portion of my contribution to the PostgreSQL open source project.

For more information on following Developer Certificate of Origin and signing off your commits, please check [here](https://github.com/babelfish-for-postgresql/babelfish_extensions/blob/main/CONTRIBUTING.md#developer-certificate-of-origin).