@forestkeeper 2023-10-23
DROP SCHEMA - the owner can drop the schema (and thereby all contained objects) even if they do not own some of the objects within the schema.
     -> Would add some reference where is this assume comes from.
DROP OWNED BY role1, role2 - drops all the objects within the current database that are owned by one of the specified roles. Any privileges granted to the given roles on objects in the current database or on shared objects (databases, tablespaces, configuration parameters) will also be revoked.
    -> Would add some reference where is this assume comes from.